aggiunta custom exception/handler per quanto riguarda la rimozione di un airbus con tratte associate

+++ src/main/java/it/prova/gestionetratte/service/AirbusServiceImpl.java
@@ -8,6 +8,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import it.prova.gestionetratte.model.Airbus;
 import it.prova.gestionetratte.repository.airbus.AirbusRepository;
+import it.prova.gestionetratte.web.api.exceptions.AirbusConTratteAssociateException;
 import it.prova.gestionetratte.web.api.exceptions.AirbusNotFoundException;
 
 @Service
@@ -52,8 +53,14 @@ public class AirbusServiceImpl implements AirbusService {
 	@Override
 	@Transactional
 	public void rimuovi(Long idToRemove) {
-		repository.findById(idToRemove)
-				.orElseThrow(() -> new AirbusNotFoundException("Airbus not found con id: " + idToRemove));
+		Airbus airbusToBeRemoved = repository.findByIdEager(idToRemove);
+		
+		if(airbusToBeRemoved == null) 
+			throw new AirbusNotFoundException("Airbus not found con id: " + idToRemove);
+		
+		if(airbusToBeRemoved.getTratte().size() > 0)
+			throw new AirbusConTratteAssociateException("Impossibile eliminare airbus: sono presenti tratte ad esso associate.");
+		
 		repository.deleteById(idToRemove);
 
 	}
